requirements.txt
autopep8==1.6.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
cycler==0.11.0
DateTime==4.4
flake8==4.0.1
flake8-broken-line==0.4.0
flake8-isort==4.1.1
flake8-plugin-utils==1.3.2
flake8-polyfill==1.0.2
flake8-return==1.1.3
fonttools==4.33.3
idna==3.3
isort==5.10.1
kiwisolver==1.4.2
matplotlib==3.5.2
mccabe==0.6.1
numpy==1.22.4
packaging==21.3
pandas==1.4.2
pep8-naming==0.12.1
Pillow==9.1.1
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
scipy==1.8.1
seaborn==0.11.2
six==1.16.0
testfixtures==6.18.5
toml==0.10.2
urllib3==1.26.9
zope.interface==5.4.0
